perf(tests): move enableCollectionYoutubeSupport off collection.service
It is the only thing in `collection.service` that needs the youtube client, and
therefore `googleapis` — which costs ~2.6s to import in a fresh process. 94 unit
test files paid that because their graph reaches collection.service; after the
move, 1 does.
